Prepare for the introduction of the new AuthFragment
