+++ app/src/main/java/com/github/versus/auth/SignInFragment.java
@@ -8,72 +8,84 @@ import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentTransaction;
 
 import com.github.versus.R;
-import com.github.versus.databinding.FragmentSignInBinding;
-import com.github.versus.db.FsUserManager;
-import com.github.versus.user.VersusUser;
+import com.github.versus.databinding.FragmentAuthBinding;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.AuthResult;
-import com.google.firebase.firestore.FirebaseFirestore;
-
-import java.util.List;
 
 /**
  * ???
  */
 public class SignInFragment extends BaseAuthFragment {
 
-    private FragmentSignInBinding binding;
+    private FragmentAuthBinding binding;
 
     @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        binding = FragmentSignInBinding.inflate(inflater);
+        binding = FragmentAuthBinding.inflate(inflater, container, false);
         return binding.getRoot();
     }
 
     @Override
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
-        registerLoginButton(binding.authSigninButton);
+
+        binding.authSignin.setOnClickListener(this::signInRequest);
+        registerLoginButton(binding.authLoginMail);
+        binding.authLoginGoogle.setOnClickListener(this::loginWithGoogleRequest);
+    }
+
+    /**
+     * ???
+     *
+     * @param view
+     */
+    private void loginWithGoogleRequest(View view) {
+        switchTo(GoogleAuthFragment.class);
+    }
+
+    /**
+     * ???
+     *
+     * @param view
+     */
+    private void signInRequest(View view) {
+        switchTo(RegisterFragment.class);
+    }
+
+    private void switchTo(Class<? extends Fragment> clz){
+        FragmentManager manager = getParentFragmentManager();
+        FragmentTransaction transaction = manager.beginTransaction();
+        transaction.replace(R.id.fragmentContainerView, clz, null);
+        transaction.commit();
     }
 
     @Override
     protected Task<AuthResult> requestAuthentication() {
-        String mailText = binding.authSigninMail.getText().toString();
-        String pwdText = binding.authSigninPwd.getText().toString();
-        Task<AuthResult> task = auth.createAccountWithMail(mailText, pwdText);
-        task.addOnSuccessListener(result -> {
-            String uid = result.getUser().getUid();
-            VersusUser.Builder builder = new VersusUser.Builder(uid);
-            // TODO HR : Link this when the UI is ready (see issue #58 in versus-app)
-            builder.setFirstName("John")
-                    .setLastName("Doe")
-                    .setUserName("johndoe")
-                    .setPhone("+41782345678")
-                    .setMail("john.doe@versus.ch")
-                    .setRating(3)
-                    .setZipCode(0)
-                    .setCity("Lausanne")
-                    .setPreferredSports(List.of());
-            new FsUserManager(FirebaseFirestore.getInstance()).insert(builder.build());
-        });
-        return task;
+        String mailText = binding.authLoginMailMail.getText().toString();
+        String pwdText = binding.authLoginMailPwd.getText().toString();
+        return auth.signInWithMail(mailText, pwdText);
     }
 
     @Override
     protected void handleSuccessfulConnection(AuthResult result) {
-        // TODO : Implement the successful result here
+
     }
 
     @Override
     protected void handleFailedConnection(Exception exception) {
-        // TODO : Implement the successful result here
+        Toast.makeText(getContext(), "handleFailedConnection",
+                Toast.LENGTH_SHORT).show();
     }
 
     @Override
     protected void handleCancelledConnection() {
-        // TODO : Implement the successful result here
+        Toast.makeText(getContext(), "handleCancelledConnection",
+                Toast.LENGTH_SHORT).show();
     }
 }
