+++ app/src/main/res/drawable-nodpi/_bg__authentication_screen_ek2_shape.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- _bg__authentication_screen_ek2 from your XD file. -->
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+	android:shape="rectangle" >
+	<solid android:color="@color/_bg__authentication_screen_ek2_color" />
+</shape>
+
