+++ app/src/main/res/drawable-nodpi/image_background_shape.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- image_background from your XD file. -->
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+	android:shape="rectangle" >
+	<corners android:radius="75dp" />
+	<solid android:color="@color/_bg__authentication_screen_ek2_color" />
+</shape>
+
