+++ app/src/main/res/drawable-nodpi/register_btn_ek1_shape.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- register_btn_ek1 from your XD file. -->
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+	android:shape="rectangle" >
+	<corners android:radius="32dp" />
+	<stroke android:color="#ffffff" android:width="8dp" />
+	<solid android:color="@color/versus_color" />
+</shape>
+
