+++ app/src/main/res/drawable-nodpi/signin_btn_ek1_shape.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- signin_btn_ek1 from your XD file. -->
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+	android:shape="rectangle" >
+	<stroke android:color="#ffffff" android:width="8dp" />
+	<solid android:color="@color/versus_color" />
+</shape>
+
