+++ app/src/main/res/layout/activity_auth.xml
@@ -8,7 +8,7 @@
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/fragmentContainerView"
-        android:name="com.github.versus.auth.AuthFragment"
+        android:name="com.github.versus.auth.SignInFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         app:layout_constraintBottom_toBottomOf="parent"
