+++ app/src/main/res/layout/fragment_sign_in.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frag_signin_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".auth.SignInFragment">
+    tools:context=".auth.RegisterFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
