+++ app/src/main/res/values/colors.xml
@@ -40,6 +40,9 @@
     <color name="rectangle_11_color">#D9D9D9</color>
     <color name="login_ek1_color">#325BC4</color>
 
+    <color name="_bg__authentication_screen_ek2_color">#FFFFFF</color>
+    <color name="versus_color">#335CC5</color>
+
     	<!-- main_chat colors -->
